requirements.txt
altair==5.3.0
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.0.1
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.63.0
grpcio-tools==1.63.0
idna==3.7
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Leonardo-Ai-SDK==5.6.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==2.0.1
packaging==24.1
pandas==2.2.2
param==2.1.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.47
protobuf==5.27.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
six==1.16.0
smmap==5.0.1
stability-sdk==0.8.6
stack-data==0.6.3
streamlit==1.36.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
wcwidth==0.2.13
